requirements.txt
djangorestframework
markdown
django-filter
psycopg2-binary